requirements.txt
# Dependencias principales
flask==2.3.3
flask-pymongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0

# Documentacion API
flasgger==0.9.7.1

# Testing
pytest==8.4.2
pytest-flask==1.3.0
pytest-mock==3.15.1